Added _sll to singlelinkedlist functions
This allows for a file to both have double linked lists in the same file as the single linked lists, without the usage of input variables depending on the type of operation to use. This way, it is harder for a user to use the incorrect function for a data structure.