Initial Commit to circular-single-linked-list fork
Added the original circular single linked list code files. One optimization to perform is to instead of all the nodes pointing to tail, create a special head node which has a reference to tail. This has the disadvantage that when adding a new head, there has to be a swap of values between the old head and the new head. Another solution is to use a Sentinel node, which is likely the best solution.