Updated circulassinglelinkedlist files
Added operations needed to work with the sentinel node. It now requires to ask for the head of the list to then iterate through the list node by node. This is not really the best solution, as it deviates from the single and double linked lists used prior, however, a compatible solution is either less efficient in memory (all nodes have tail pointers), or less efficient in time (to insert/remove  head/tail, the traversal through the hole list has to be done).