Initial Implementation of circular double linked list
Initial Implementation of circular double linked list. TODO: the insert/delete into n will not work, this is because since the head previous does not point to null, it is impossible to know if the node is a head or not. This can be corrected by only allowing the head to be sent as a node input, or to not care what the head of the list is (that is in a circular list there is really no head). This requires the user not to free/change the head of the list, if he has an auxiliary pointer to that head in a different variable.