Update set.c
Added resize function, set insert, set erase and set contains. The hash function is still not implemented, and might have to be passed by the user, since it highly depends on the data sent (for integers its easy, for strings in a bit more challenging, for complex structs it would be impossible). Still not tested.