Create set_test.c
Initial commit of set_test file. The particular function added (single_set_simple_test) checks for the existence of elements and how many are in the set. Furthermore, it also adds a few elements and checks for their presence then removes them and checks again. Finally, it frees the entire set.